@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "lint": "nx workspace-lint && ng lint",
@@ -12,16 +12,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.11",
-    "@angular/cdk": "13.3.8",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
-    "@nrwl/angular": "14.1.9",
+    "@angular/animations": "14.0.1",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.1",
+    "@angular/compiler": "14.0.1",
+    "@angular/core": "14.0.1",
+    "@angular/forms": "14.0.1",
+    "@angular/platform-browser": "14.0.1",
+    "@angular/platform-browser-dynamic": "14.0.1",
+    "@angular/router": "14.0.1",
+    "@ngrx/component-store": "13.2.0",
+    "@ngrx/effects": "13.2.0",
+    "@ngrx/entity": "13.2.0",
+    "@ngrx/router-store": "13.2.0",
+    "@ngrx/store": "13.2.0",
+    "@nrwl/angular": "14.2.4",
     "animate.css": "^4.1.1",
     "i18n-iso-countries": "^7.4.0",
     "normalize.css": "^8.0.1",
@@ -34,33 +39,35 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.7",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "13.3.7",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
-    "@nrwl/cli": "14.1.9",
-    "@nrwl/cypress": "14.1.9",
-    "@nrwl/eslint-plugin-nx": "14.1.9",
-    "@nrwl/jest": "14.1.9",
-    "@nrwl/linter": "14.1.9",
-    "@nrwl/workspace": "14.1.9",
+    "@angular-devkit/build-angular": "14.0.1",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.1",
+    "@angular/language-service": "14.0.1",
+    "@ngrx/schematics": "13.2.0",
+    "@ngrx/store-devtools": "13.2.0",
+    "@nrwl/cli": "14.2.4",
+    "@nrwl/cypress": "14.2.4",
+    "@nrwl/eslint-plugin-nx": "14.2.4",
+    "@nrwl/jest": "14.2.4",
+    "@nrwl/linter": "14.2.4",
+    "@nrwl/workspace": "14.2.4",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "5.18.0",
-    "@typescript-eslint/parser": "5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "cypress": "^9.1.0",
-    "eslint": "8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
+    "jasmine-marbles": "~0.9.1",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "prettier": "^2.5.1",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.2.4",
+    "prettier": "2.6.2",
     "ts-jest": "27.1.4",
-    "typescript": "4.6.4",
-    "nx": "14.1.9"
+    "typescript": "4.7.3"
   }
 }
-
