@@ -12,8 +12,12 @@ import { AccordionModule } from 'primeng/accordion'
 import { NavComponent } from './shared/nav/nav.component'
 import { ProductsModule } from '@my-qart/products'
 import { UiModule } from '@frontend/ui'
-import { HttpClientModule } from '@angular/common/http'
+import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http'
 import { OrdersModule } from '@my-qart/orders'
+import {JwtInterceptor, UsersModule} from "@my-qart/users";
+import {MessageService} from "primeng/api";
+import {StoreModule} from "@ngrx/store";
+import {EffectsModule} from "@ngrx/effects";
 
 const routes: Routes = [{ path: '', component: HomePageComponent }]
 
@@ -31,12 +35,17 @@ const routes: Routes = [{ path: '', component: HomePageComponent }]
     BrowserAnimationsModule,
     RouterModule.forRoot(routes),
     HttpClientModule,
+    EffectsModule.forRoot([]),
+    StoreModule.forRoot({}),
     ProductsModule,
     UiModule,
     AccordionModule,
-    OrdersModule
+    OrdersModule,
+    UsersModule
   ],
-  providers: [],
+  providers: [MessageService, {
+    provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true
+  }],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
