@@ -1,9 +1,10 @@
-import { User } from '../models/user'
-import { HttpClient } from '@angular/common/http'
-import { Injectable } from '@angular/core'
-import { environment } from '@env/environment'
-import { map, Observable } from 'rxjs'
+import {User} from '../models/user'
+import {HttpClient} from '@angular/common/http'
+import {Injectable} from '@angular/core'
+import {environment} from '@env/environment'
+import {map, Observable} from 'rxjs'
 import * as countriesLib from 'i18n-iso-countries'
+import {UsersFacade} from "../+state/users.facade";
 
 declare const require: any
 
@@ -13,7 +14,7 @@ declare const require: any
 export class UsersService {
   apiURLUsers = environment.apiURL + '/users'
 
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient, private usersFacade: UsersFacade) {
     countriesLib.registerLocale(require('i18n-iso-countries/langs/en.json'))
   }
 
@@ -39,7 +40,7 @@ export class UsersService {
 
   getCountries(): { value: string; label: string }[] {
     return Object.entries(
-      countriesLib.getNames('en', { select: 'official' })
+      countriesLib.getNames('en', {select: 'official'})
     ).map((entry) => {
       return {
         value: entry[0],
@@ -56,5 +57,17 @@ export class UsersService {
     return this.http
       .get<number>(`${this.apiURLUsers}/get/count`)
       .pipe(map((objectValue: any) => objectValue.userCount))
+  }
+
+  initAppSession() {
+    this.usersFacade.buildUserSession()
+  }
+
+  observeCurrentUser(){
+    return this.usersFacade.currentUser$
+  }
+
+  isCurrentUserAuthenticated(){
+    return this.usersFacade.isAuthenticated$
   }
 }
